requirement.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.13
albumentations==1.4.12
annotated-types==0.7.0
astunparse==1.6.3
-e git+https://github.com/megvii-model/HINet@4e7231543090e6280d03fac22b3bb6869a25dfad#egg=basicsr
build==1.2.1
cached-property==1.5.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chex==0.1.7
clu==0.0.12
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
dm-tree==0.1.8
einops==0.8.0
etils==1.3.0
eval_type_backport==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
flax==0.7.2
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.2
h5py==3.11.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.2.0
importlib_resources==6.4.0
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.4
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
Levenshtein==0.25.1
libclang==18.1.1
lmdb==1.5.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
maxim @ file:///home/ayakashikitsune/Documents/ProgramingProjects/maximulation/maxim
mdurl==0.1.2
ml-collections==0.1.0
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.8
munch==2.5.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optax==0.1.8
orbax-checkpoint==0.2.3
packaging==24.1
pandas==2.0.3
pillow==10.4.0
platformdirs==4.2.2
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.9.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.13.1
tb-nightly==2.14.0a20230808
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorstore==0.1.45
termcolor==2.4.0
tifffile==2023.7.10
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.43.0
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
zipp==3.19.2
